4.22

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,9 @@
 "use strict";
 const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
+const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+const config = require("../utils/config");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
@@ -9,8 +12,12 @@ blogsRouter.get("/", async (request, response) => {
 
 blogsRouter.post("/", async (request, response) => {
   const body = request.body;
-
-  const user = await User.findById(body.userId);
+  const token = request.token;
+  const decodedToken = jwt.verify(token, config.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
 
   const blog = new Blog({ ...body, user: user._id });
 
@@ -22,22 +29,46 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
+  const token = request.token;
+  const decodedToken = jwt.verify(token, config.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const blog = await Blog.findById(request.params.id);
+  const user = await User.findById(decodedToken.id);
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: "incorrect user" });
+  }
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
 
 blogsRouter.put("/:id", async (request, response) => {
+  const token = request.token;
+  const decodedToken = jwt.verify(token, config.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const blog = await Blog.findById(request.params.id);
+  const user = await User.findById(decodedToken.id);
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: "incorrect user" });
+  }
   const body = request.body;
-  const blog = {
+  const updateBlogInfo = {
     title: body.title,
     author: body.author,
     likes: body.likes,
     url: body.url,
   };
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
-    new: true,
-  });
-  response.json(updatedBlog);
+  const updatedBlog = await Blog.findByIdAndUpdate(
+    request.params.id,
+    updateBlogInfo,
+    {
+      new: true,
+    }
+  );
+  return response.json(updatedBlog);
 });
 
 module.exports = blogsRouter;
